Authorization of projects CRUD handled.

+++ projects/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 
 from .models import Project
 from .forms import ProjectForm
@@ -17,12 +18,15 @@ def project(request, pk):
     return render(request, 'projects/project.html', context)
 
 
+@login_required(login_url='login')
 def create_project(request):
     form = ProjectForm()
 
     if request.method == "POST":
         form = ProjectForm(request.POST, request.FILES)
         if form.is_valid:
+            project = form.save(commit=False)
+            project.owner = request.user
             form.save()
             return redirect('/')
 
@@ -30,22 +34,30 @@ def create_project(request):
     return render(request, 'projects/create-project.html', context)
 
 
+@login_required(login_url='login')
 def update_project(request, pk):
     project = Project.objects.get(id=pk)
     form = ProjectForm(instance=project)
+
+    if request.user.profile != project.owner:
+        return render(request, 'unauthorized.html')
 
     if request.method == "POST":
         form = ProjectForm(request.POST, request.FILES, instance=project)
         if form.is_valid:
             form.save()
-            return redirect('/')
+            return redirect('project', project.id)
 
     context = {'form':form}
     return render(request, 'projects/create-project.html', context)
 
 
+@login_required(login_url='login')
 def delete_project(request, pk):
     project = Project.objects.get(id=pk)
+
+    if request.user.profile != project.owner:
+        return render(request, 'unauthorized.html')
 
     if request.method == "POST":
         project.delete()
